@@ -13,42 +13,34 @@
 		"test": "jest"
 	},
 	"dependencies": {
-		"@apollo/react-hooks": "^3.1.3",
-		"@react-native-community/netinfo": "~3.2.1",
-		"apollo": "^2.21.1",
-		"apollo-boost": "^0.4.4",
-		"apollo-cache-inmemory": "^1.6.3",
-		"apollo-client": "^2.6.4",
-		"apollo-link-http": "^1.5.16",
-		"axios": "^0.18.0",
-		"expo": "^35.0.0",
-		"expo-ads-admob": "~7.0.0",
-		"expo-asset": "~7.0.0",
-		"expo-font": "~7.0.0",
-		"expo-gl": "~7.0.0",
-		"expo-gl-cpp": "~7.0.0",
-		"expo-keep-awake": "~7.0.0",
-		"expo-secure-store": "^7.0.0",
-		"expo-three": "^5.2.0",
-		"expokit": "32.0.7",
-		"firebase": "^5.5.4",
-		"graphql": "^14.5.8",
-		"graphql-tag": "^2.10.1",
-		"react": "16.8.3",
-		"react-apollo": "^3.1.3",
-		"react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
-		"react-native-scripts": "2.0.1",
+		"expo": "^39.0.0",
+		"expo-ads-admob": "~8.3.0",
+		"expo-asset": "~8.2.0",
+		"expo-font": "~8.3.0",
+		"expo-gl": "^9.1.1",
+		"expo-gl-cpp": "~9.1.1",
+		"expo-secure-store": "~9.2.0",
+		"expo-three": "^5.5.1",
+		"expo-updates": "~0.3.3",
+		"firebase": "7.9.0",
+		"react": "16.13.1",
+		"react-native": "https://github.com/expo/react-native/archive/sdk-39.0.3.tar.gz",
+		"react-native-gesture-handler": "^1.9.0",
+		"react-native-reanimated": "^1.13.2",
+		"react-native-screens": "^2.15.0",
 		"react-native-vector-icons": "^6.0.2",
-		"react-navigation": "^2.17.0",
+		"react-navigation": "^4.4.3",
+		"react-navigation-stack": "^2.10.2",
 		"react-redux": "^5.0.7",
 		"redux": "^4.0.0",
 		"simplex-noise": "^2.4.0",
-		"three": "^0.110.0",
-		"three-bas": "^2.10.1"
+		"three": "^0.123.0",
+		"three-bas": "^2.12.0"
 	},
 	"devDependencies": {
+		"@react-native-community/masked-view": "^0.1.10",
+		"@react-native-community/slider": "^3.0.3",
 		"babel": "^6.23.0",
-		"babel-plugin-module-alias": "^1.6.0",
-		"schedule": "^0.4.0"
+		"babel-plugin-module-alias": "^1.6.0"
 	}
 }
