@@ -1,8 +1,9 @@
-import Expo, { Updates } from "expo";
+import Expo from "expo";
+import * as Updates from 'expo-updates'
 const actions = require("./actions");
 import store from "./store";
 import * as SecureStore from "expo-secure-store";
-import * as firebase from "firebase";
+// import * as firebase from "firebase";
 import { AdMobInterstitial } from "expo-ads-admob";
 import AdMob from "../constants/AdMob";
 import FIREBASE_CONSTANTS from "../constants/firebase";
@@ -18,10 +19,10 @@ export default (dispatch) => () => {
 
 	const initialize = async () => {
 		// Initialize Firebase
-		if (!store.getState().initialized) {
-			firebase.initializeApp(FIREBASE_CONSTANTS);
-			set("/initialized", true);
-		}
+		// if (!store.getState().initialized) {
+		// 	firebase.initializeApp(FIREBASE_CONSTANTS);
+		// 	set("/initialized", true);
+		// }
 
 		// AdMobInterstitial.setAdUnitID(AdMob.interstitial); // Test ID, Replace with your-admob-unit-id
 		// // AdMobInterstitial.setTestDeviceID('EMULATOR');
@@ -38,46 +39,46 @@ export default (dispatch) => () => {
 		// check if server has info?
 
 		// check auth login state changes
-		firebase.auth().onAuthStateChanged(async (user) => {
-			// console.log("!!onAuthStateChanged", user != null);
-			if (user != null) {
-				// get the user's local highscore to compare to the server highscore, take whatever is higher
-				// server is what other users will see but secure store lets users play offline and keep their highscore...
-				// if they play offline they will compete against themselves, but when they go online and they play their new highscore will in fact be updated online
-				// TODO: add this updating online/offline to the ABOUT page
-				const userHsFromSecureStore = await getSecureStoreHighScore();
+		// firebase.auth().onAuthStateChanged(async (user) => {
+		// 	// console.log("!!onAuthStateChanged", user != null);
+		// 	if (user != null) {
+		// 		// get the user's local highscore to compare to the server highscore, take whatever is higher
+		// 		// server is what other users will see but secure store lets users play offline and keep their highscore...
+		// 		// if they play offline they will compete against themselves, but when they go online and they play their new highscore will in fact be updated online
+		// 		// TODO: add this updating online/offline to the ABOUT page
+		// 		const userHsFromSecureStore = await getSecureStoreHighScore();
 
-				// get the current user"s highscore
-				firebase
-					.database()
-					.ref("users/" + user.uid)
-					.once("value", (snapshot) => {
-						let hs = snapshot.val() && snapshot.val().highscore;
-						if (!hs || hs < userHsFromSecureStore) {
-							// if server highscore is less than secure store hardware highscore, choose the secure store hardware highscore
-							hs = userHsFromSecureStore;
-							// since server is not as high as local secure store, also update the server
-							const user = firebase.auth().currentUser;
-							updateUserOnDb({
-								uid: user.uid,
-								highscore: hs,
-							});
-						}
-						// set the highscore to whatever the server or secure store hardware value is
-						setUserHighscore(hs);
-					});
+		// 		// get the current user"s highscore
+		// 		firebase
+		// 			.database()
+		// 			.ref("users/" + user.uid)
+		// 			.once("value", (snapshot) => {
+		// 				let hs = snapshot.val() && snapshot.val().highscore;
+		// 				if (!hs || hs < userHsFromSecureStore) {
+		// 					// if server highscore is less than secure store hardware highscore, choose the secure store hardware highscore
+		// 					hs = userHsFromSecureStore;
+		// 					// since server is not as high as local secure store, also update the server
+		// 					const user = firebase.auth().currentUser;
+		// 					updateUserOnDb({
+		// 						uid: user.uid,
+		// 						highscore: hs,
+		// 					});
+		// 				}
+		// 				// set the highscore to whatever the server or secure store hardware value is
+		// 				setUserHighscore(hs);
+		// 			});
 
-				// since user exists, store user data in the reducer
+		// 		// since user exists, store user data in the reducer
 
-				setUserDataLocal(user);
-			} else {
-				const hs = await getSecureStoreHighScore();
-				setUserHighscore(hs);
-			}
+		// 		setUserDataLocal(user);
+		// 	} else {
+				// const hs = await getSecureStoreHighScore();
+				// setUserHighscore(hs);
+			// }
 			// Do other things
-		});
+		// });
 
-		getHighscores();
+		// getHighscores();
 	};
 
 	const toggleModal = (action) => {
@@ -132,21 +133,21 @@ export default (dispatch) => () => {
 
 	const getHighscores = () => {
 		// a promise is used here to create a loader
-		return new Promise((resolve, reject) => {
-			// get all users" highscores
-			var query = firebase.database().ref("users");
-			query
-				.orderByChild("highscore")
-				.limitToLast(10)
-				.on("value", (snapshot) => {
-					let highscores = [];
-					snapshot.forEach((childSnapshot) => {
-						highscores.push(childSnapshot.val());
-					});
-					set("/highscores", highscores);
-					resolve(highscores);
-				});
-		});
+		// return new Promise((resolve, reject) => {
+		// 	// get all users" highscores
+		// 	var query = firebase.database().ref("users");
+		// 	query
+		// 		.orderByChild("highscore")
+		// 		.limitToLast(10)
+		// 		.on("value", (snapshot) => {
+		// 			let highscores = [];
+		// 			snapshot.forEach((childSnapshot) => {
+		// 				highscores.push(childSnapshot.val());
+		// 			});
+		// 			set("/highscores", highscores);
+		// 			resolve(highscores);
+		// 		});
+		// });
 	};
 
 	const getSecureStoreHighScore = async () => {
@@ -187,10 +188,10 @@ export default (dispatch) => () => {
 
 	// update the DB user data (happens after each login to ensure redundancy)
 	const updateUserOnDb = (obj) => {
-		firebase
-			.database()
-			.ref("users/" + obj.uid)
-			.update(obj);
+		// firebase
+		// 	.database()
+		// 	.ref("users/" + obj.uid)
+		// 	.update(obj);
 	};
 
 	const setFirstTime = () => {
@@ -231,32 +232,32 @@ export default (dispatch) => () => {
 			console.log("done clearing local data ");
 		});
 		clearUserDataLocal();
-		firebase.auth().signOut();
+		// firebase.auth().signOut();
 	};
 	const signUp = (email, password, displayName) => {
-		firebase
-			.auth()
-			.createUserWithEmailAndPassword(email, password)
-			.then(
-				({ user }) => {
-					updateUserOnDb({
-						uid: user.uid,
-						displayName: displayName,
-						created: Date.now(),
-						email: user.email,
-						highscore: 0 || store.getState().deltaTime,
-					});
-				},
-				(error) => {
-					console.log("ERROR");
-				},
-			);
+		// firebase
+		// 	.auth()
+		// 	.createUserWithEmailAndPassword(email, password)
+		// 	.then(
+		// 		({ user }) => {
+		// 			updateUserOnDb({
+		// 				uid: user.uid,
+		// 				displayName: displayName,
+		// 				created: Date.now(),
+		// 				email: user.email,
+		// 				highscore: 0 || store.getState().deltaTime,
+		// 			});
+		// 		},
+		// 		(error) => {
+		// 			console.log("ERROR");
+		// 		},
+		// 	);
 	};
 	const login = (email, password) => {
-		return firebase
-			.auth()
-			.signInWithEmailAndPassword(email, password)
-			.catch((error) => error);
+		// return firebase
+		// 	.auth()
+		// 	.signInWithEmailAndPassword(email, password)
+		// 	.catch((error) => error);
 	};
 	const checkUserUid = async () => {
 		let uid = await getUserUid();
