@@ -1,14 +1,13 @@
 import Level from "./Level";
 import ExpoTHREE, { THREE } from "expo-three"; // 3.0.0-alpha.4
-import Particles from "../Visualizations/Particles";
 
 export default class VortexLevel extends Level {
 	constructor(options) {
 		super(options);
 		console.log("VortexLevel constructor");
 		
-		const texture = await ExpoTHREE.loadTextureAsync({
-			asset: require("../../assets/images/stars.jpg"),
+		const texture = ExpoTHREE.loadTextureAsync({
+			asset: require("../../../assets/images/stars.jpg"),
 		});
 		this.visualization = new Vortex(texture);
 	}
